+++ examples/hitechnic-pir-test1.c
@@ -24,7 +24,7 @@
 #include "hitechnic-pir.h"
 
 // Draw a little stick figure
-void drawStickFigure(int offset)
+void drawStickFigure(short offset)
 {
   eraseRect(0, 0, 99, 50);
   drawCircle(offset + 43, 40, 15);
@@ -35,9 +35,9 @@ void drawStickFigure(int offset)
 }
 
 task main () {
-  int reading = 0;
-  int x = 0;
-  int deadband = HTPIR_DEFAULT_DEADBAND;
+  short reading = 0;
+  short x = 0;
+  short deadband = HTPIR_DEFAULT_DEADBAND;
 
   displayCenteredTextLine(0, "HiTechnic");
   displayCenteredBigTextLine(1, "PIR");
