+++ examples/lego-light-SMUX-test2.c
@@ -35,8 +35,8 @@
 const tMUXSensor LEGOLS = msensor_S1_4;
 
 task main () {
-  int raw = 0;
-  int nrm = 0;
+  short raw = 0;
+  short nrm = 0;
   // Get control over the buttons
   nNxtButtonTask  = -2;
