+++ examples/lego-light-test1.c
@@ -24,8 +24,8 @@
 #include "lego-light.h"
 
 task main() {
-  int raw = 0;
-  int nrm = 0;
+  short raw = 0;
+  short nrm = 0;
   bool active = true;
   LSsetActive(LEGOLS);
